Update to de_DE.md
Started translating the actual Text. I had to make some grammatical changes to make it read more naturally but generally I tried to stay as close to the original as possible. Still, a second opinion would be great. I also changed the asterisks for gendering to underscores so as to hopefully prevent unwanted md-formatting in the future. :) Feedback and changes are always appreciated of course.